qefd

+++ src/main/java/io/mysticalshadow/easyservermanager/commands/CMD_World.java
@@ -27,7 +27,6 @@ public class CMD_World implements CommandExecutor {
     public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
         if (sender instanceof Player) {
             Player p = (Player) sender;
-            if (p.hasPermission(plugin.PermsWorld) || p.hasPermission(plugin.PermSternchen)) {
                 if (args.length == 0) {
                     p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world create <world> <type>");
                     p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp <world>");
@@ -36,31 +35,36 @@ public class CMD_World implements CommandExecutor {
                 }
                 if (args.length == 3) {
                     if (args[0].equalsIgnoreCase("create")) {
-                        String worldName = args[1];
-                        World w = Bukkit.getWorld(worldName);
-                        if (w == null) {
-                            String worldCreated = plugin.WorldCreatedMSG;
-                            worldCreated = worldCreated.replace("%world%", worldName);
-                            if (args[2].equalsIgnoreCase("normal")) {
-                                Bukkit.createWorld(WorldCreator.name(worldName).type(WorldType.NORMAL));
-                                p.sendMessage(plugin.Prefix + worldCreated);
-                                p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp " + worldName);
-                                return true;
-                            } else if (args[2].equalsIgnoreCase("flat")) {
-                                Bukkit.createWorld(WorldCreator.name(worldName).type(WorldType.FLAT));
-                                p.sendMessage(plugin.Prefix + worldCreated);
-                                p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp " + worldName);
-                                return true;
-                            } else if (args[2].equalsIgnoreCase("large")) {
-                                Bukkit.createWorld(WorldCreator.name(worldName).type(WorldType.LARGE_BIOMES));
-                                p.sendMessage(plugin.Prefix + worldCreated);
-                                p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp " + worldName);
-                                return true;
+                        if (p.hasPermission(plugin.PermsWorldCreate) || p.hasPermission(plugin.PermSternchen)) {
+                            String worldName = args[1];
+                            World w = Bukkit.getWorld(worldName);
+                            if (w == null) {
+                                String worldCreated = plugin.WorldCreatedMSG;
+                                worldCreated = worldCreated.replace("%world%", worldName);
+                                if (args[2].equalsIgnoreCase("normal")) {
+                                    Bukkit.createWorld(WorldCreator.name(worldName).type(WorldType.NORMAL));
+                                    p.sendMessage(plugin.Prefix + worldCreated);
+                                    p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp " + worldName);
+                                    return true;
+                                } else if (args[2].equalsIgnoreCase("flat")) {
+                                    Bukkit.createWorld(WorldCreator.name(worldName).type(WorldType.FLAT));
+                                    p.sendMessage(plugin.Prefix + worldCreated);
+                                    p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp " + worldName);
+                                    return true;
+                                } else if (args[2].equalsIgnoreCase("large")) {
+                                    Bukkit.createWorld(WorldCreator.name(worldName).type(WorldType.LARGE_BIOMES));
+                                    p.sendMessage(plugin.Prefix + worldCreated);
+                                    p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world tp " + worldName);
+                                    return true;
+                                }
+                            } else {
+                                String worldAlreadyExist = plugin.WorldAlreadyExistMSG;
+                                worldAlreadyExist = worldAlreadyExist.replace("%world%", worldName);
+                                p.sendMessage(plugin.Prefix + worldAlreadyExist);
                             }
                         } else {
-                            String worldAlreadyExist = plugin.WorldAlreadyExistMSG;
-                            worldAlreadyExist = worldAlreadyExist.replace("%world%", worldName);
-                            p.sendMessage(plugin.Prefix + worldAlreadyExist);
+                            p.sendMessage(plugin.Prefix + plugin.NoPermMessage);
+                            return true;
                         }
                     }
                 } else if (args[0].equalsIgnoreCase("create")) {
@@ -69,18 +73,23 @@ public class CMD_World implements CommandExecutor {
                 }
                 if (args.length == 2) {
                     if (args[0].equalsIgnoreCase("tp")) {
-                        String worldName = args[1];
-                        World w = Bukkit.getWorld(worldName);
-                        if (w != null) {
-                            p.teleport(Bukkit.getWorld(worldName).getSpawnLocation());
-                            String worldTeleported = plugin.YourTeleportToTheWorldMSG;
-                            worldTeleported = worldTeleported.replace("%world%", worldName);
-                            p.sendMessage(plugin.Prefix + worldTeleported);
-                            return true;
+                        if (p.hasPermission(plugin.PermsWorldTP) || p.hasPermission(plugin.PermSternchen)) {
+                            String worldName = args[1];
+                            World w = Bukkit.getWorld(worldName);
+                            if (w != null) {
+                                p.teleport(Bukkit.getWorld(worldName).getSpawnLocation());
+                                String worldTeleported = plugin.YourTeleportToTheWorldMSG;
+                                worldTeleported = worldTeleported.replace("%world%", worldName);
+                                p.sendMessage(plugin.Prefix + worldTeleported);
+                                return true;
+                            } else {
+                                String worldNotExist = plugin.WorldDoesNotExistMSG;
+                                worldNotExist = worldNotExist.replace("%world%", worldName);
+                                p.sendMessage(plugin.Prefix + worldNotExist);
+                            }
                         } else {
-                            String worldNotExist = plugin.WorldDoesNotExistMSG;
-                            worldNotExist = worldNotExist.replace("%world%", worldName);
-                            p.sendMessage(plugin.Prefix + worldNotExist);
+                            p.sendMessage(plugin.Prefix + plugin.NoPermMessage);
+                            return true;
                         }
                     }
                 } else if (args[0].equalsIgnoreCase("tp")) {
@@ -89,28 +98,30 @@ public class CMD_World implements CommandExecutor {
                 }
                 if (args.length == 2) {
                     if (args[0].equalsIgnoreCase("remove")) {
-                        String worldName = args[1];
-                        World w = Bukkit.getWorld(worldName);
-                        if (w != null) {
-                            Bukkit.unloadWorld(worldName, true);
-                            deleteFolder(w.getWorldFolder().toPath());
-                            String worldRemoved = plugin.YourRemovedTheWorldMSG;
-                            worldRemoved = worldRemoved.replace("%world%", worldName);
-                            p.sendMessage(plugin.Prefix + worldRemoved);
-                            return true;
+                        if (p.hasPermission(plugin.PermsWorldRemove) || p.hasPermission(plugin.PermSternchen)) {
+                            String worldName = args[1];
+                            World w = Bukkit.getWorld(worldName);
+                            if (w != null) {
+                                Bukkit.unloadWorld(worldName, true);
+                                deleteFolder(w.getWorldFolder().toPath());
+                                String worldRemoved = plugin.YourRemovedTheWorldMSG;
+                                worldRemoved = worldRemoved.replace("%world%", worldName);
+                                p.sendMessage(plugin.Prefix + worldRemoved);
+                                return true;
+                            } else {
+                                String worldNotExist = plugin.WorldDoesNotExistMSG;
+                                worldNotExist = worldNotExist.replace("%world%", worldName);
+                                p.sendMessage(plugin.Prefix + worldNotExist);
+                            }
                         } else {
-                            String worldNotExist = plugin.WorldDoesNotExistMSG;
-                            worldNotExist = worldNotExist.replace("%world%", worldName);
-                            p.sendMessage(plugin.Prefix + worldNotExist);
+                            p.sendMessage(plugin.Prefix + plugin.NoPermMessage);
+                            return true;
                         }
                     }
                 } else if (args[0].equalsIgnoreCase("remove")) {
                     p.sendMessage(plugin.Prefix + plugin.UseCommandMSG + "world remove <world>");
                     return true;
                 }
-            } else {
-                p.sendMessage(plugin.Prefix + plugin.NoPermMessage);
-            }
         }
         return false;
     }
